# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DIRO Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  counselor: 'counselor',
  user: 'user',
} as const;

export interface User {
  id: string;
  discord_id: string;
  username: string;
  /** @nullable */
  display_name?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  email?: string | null;
  role: UserRole;
  created_at: string;
  /** @nullable */
  last_login?: string | null;
}

export interface UserUpdate {
  display_name?: string;
  avatar?: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  consulting: 'consulting',
  building: 'building',
  payment_pending: 'payment_pending',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type OrderAtmosphere = typeof OrderAtmosphere[keyof typeof OrderAtmosphere];


export const OrderAtmosphere = {
  gaming: 'gaming',
  community: 'community',
  corporate: 'corporate',
  social: 'social',
  streaming: 'streaming',
  education: 'education',
  other: 'other',
} as const;

export interface Order {
  id: string;
  order_number: string;
  user_id: string;
  /** @nullable */
  counselor_id?: string | null;
  status: OrderStatus;
  server_name: string;
  /** @nullable */
  server_description?: string | null;
  atmosphere: OrderAtmosphere;
  /** @nullable */
  category_count?: number | null;
  /** @nullable */
  text_channel_count?: number | null;
  /** @nullable */
  voice_channel_count?: number | null;
  /** @nullable */
  desired_roles?: string | null;
  /** @nullable */
  desired_permissions?: string | null;
  /** @nullable */
  desired_features?: string | null;
  budget: number;
  /** @nullable */
  additional_notes?: string | null;
  created_at: string;
  updated_at: string;
  /** @nullable */
  user_username?: string | null;
  /** @nullable */
  user_avatar?: string | null;
  /** @nullable */
  user_display_name?: string | null;
  /** @nullable */
  counselor_username?: string | null;
  /** @nullable */
  counselor_avatar?: string | null;
}

export type OrderInputAtmosphere = typeof OrderInputAtmosphere[keyof typeof OrderInputAtmosphere];


export const OrderInputAtmosphere = {
  gaming: 'gaming',
  community: 'community',
  corporate: 'corporate',
  social: 'social',
  streaming: 'streaming',
  education: 'education',
  other: 'other',
} as const;

export interface OrderInput {
  server_name: string;
  server_description?: string;
  atmosphere: OrderInputAtmosphere;
  category_count?: number;
  text_channel_count?: number;
  voice_channel_count?: number;
  desired_roles?: string;
  desired_permissions?: string;
  desired_features?: string;
  budget: number;
  additional_notes?: string;
}

export type OrderUpdateStatus = typeof OrderUpdateStatus[keyof typeof OrderUpdateStatus];


export const OrderUpdateStatus = {
  pending: 'pending',
  consulting: 'consulting',
  building: 'building',
  payment_pending: 'payment_pending',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderUpdate {
  status?: OrderUpdateStatus;
  counselor_id?: string;
}

export interface OrderStats {
  message_count: number;
  days_elapsed: number;
  progress_percent: number;
  /** @nullable */
  last_activity?: string | null;
}

export type MessageType = typeof MessageType[keyof typeof MessageType];


export const MessageType = {
  text: 'text',
  image: 'image',
  file: 'file',
  system: 'system',
  payment: 'payment',
  preview: 'preview',
} as const;

export interface Message {
  id: string;
  order_id: string;
  sender_id: string;
  content: string;
  type: MessageType;
  /** @nullable */
  metadata_json?: string | null;
  is_edited?: boolean;
  created_at: string;
  /** @nullable */
  updated_at?: string | null;
  /** @nullable */
  sender_username?: string | null;
  /** @nullable */
  sender_avatar?: string | null;
  /** @nullable */
  sender_display_name?: string | null;
  /** @nullable */
  sender_role?: string | null;
}

export type MessageInputType = typeof MessageInputType[keyof typeof MessageInputType];


export const MessageInputType = {
  text: 'text',
  image: 'image',
  file: 'file',
  system: 'system',
  payment: 'payment',
  preview: 'preview',
} as const;

export interface MessageInput {
  content: string;
  type: MessageInputType;
  metadata_json?: string;
}

export interface MessageUpdate {
  content: string;
}

export interface ServerProject {
  id: string;
  order_id: string;
  config_json: string;
  /** @nullable */
  history_json?: string | null;
  created_at: string;
  /** @nullable */
  updated_at?: string | null;
}

export interface ServerProjectUpdate {
  config_json: string;
}

export interface PreviewResult {
  success: boolean;
  message_id: string;
}

export type PaymentRequestStatus = typeof PaymentRequestStatus[keyof typeof PaymentRequestStatus];


export const PaymentRequestStatus = {
  pending: 'pending',
  paid: 'paid',
  cancelled: 'cancelled',
} as const;

export interface PaymentRequest {
  id: string;
  order_id: string;
  counselor_id: string;
  amount: number;
  status: PaymentRequestStatus;
  deeplink: string;
  created_at: string;
  /** @nullable */
  updated_at?: string | null;
}

export interface PaymentRequestInput {
  order_id: string;
  amount: number;
}

export type PaymentRequestUpdateStatus = typeof PaymentRequestUpdateStatus[keyof typeof PaymentRequestUpdateStatus];


export const PaymentRequestUpdateStatus = {
  pending: 'pending',
  paid: 'paid',
  cancelled: 'cancelled',
} as const;

export interface PaymentRequestUpdate {
  status: PaymentRequestUpdateStatus;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  new_message: 'new_message',
  new_order: 'new_order',
  status_change: 'status_change',
  preview_sent: 'preview_sent',
  payment_request: 'payment_request',
} as const;

export interface Notification {
  id: string;
  user_id: string;
  type: NotificationType;
  title: string;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  reference_id?: string | null;
  is_read: boolean;
  created_at: string;
}

export interface DiscordVerifyInput {
  server_id: string;
}

export interface DiscordVerifyResult {
  success: boolean;
  in_server: boolean;
  /** @nullable */
  server_name?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface DiscordApplyInput {
  order_id: string;
  server_id: string;
}

export interface DiscordApplyResult {
  success: boolean;
  applied_items?: string[];
  /** @nullable */
  error?: string | null;
}

export interface AdminStats {
  total_orders: number;
  total_users: number;
  total_counselors: number;
  pending_orders?: number;
  consulting_orders?: number;
  building_orders?: number;
  completed_orders?: number;
  cancelled_orders?: number;
  revenue_total: number;
  orders_this_week: number;
  active_chats: number;
}

export type AdminUserUpdateRole = typeof AdminUserUpdateRole[keyof typeof AdminUserUpdateRole];


export const AdminUserUpdateRole = {
  admin: 'admin',
  counselor: 'counselor',
  user: 'user',
} as const;

export interface AdminUserUpdate {
  role: AdminUserUpdateRole;
}

export interface Announcement {
  id: string;
  title: string;
  content: string;
  created_at: string;
}

export interface AnnouncementInput {
  title: string;
  content: string;
}

export type GetOrdersParams = {
status?: string;
limit?: number;
offset?: number;
};

export type GetNotificationsParams = {
unread_only?: boolean;
};

export type GetAdminUsersParams = {
role?: string;
limit?: number;
offset?: number;
};

export type GetAdminOrdersParams = {
status?: string;
limit?: number;
offset?: number;
};