# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DIRO Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, discord_id, username, role, created_at]
      properties:
        id:
          type: string
        discord_id:
          type: string
        username:
          type: string
        display_name:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, counselor, user]
        created_at:
          type: string
        last_login:
          type: ["string", "null"]
      additionalProperties: false
    UserUpdate:
      type: object
      properties:
        display_name:
          type: string
        avatar:
          type: string
      additionalProperties: false
...
    AdminUserUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, counselor, user]
      additionalProperties: false
...